requirements.txt
black==22.3.0
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyYAML==6.0
tomli==2.0.1
typing_extensions==4.2.0
